# Conflicts:
#	.gradle/3.3/taskArtifacts/fileHashes.bin
#	.gradle/3.3/taskArtifacts/fileSnapshots.bin
#	.gradle/3.3/taskArtifacts/taskArtifacts.bin
#	.gradle/3.3/taskArtifacts/taskArtifacts.lock
#
.gradle/3.3/tasks/_app_compileDebugJavaWithJavac/localClassSetAnalysis/l
ocalClassSetAnalysis.lock
#
.gradle/3.3/tasks/_app_compileDebugJavaWithJavac/localJarClasspathSnapsh
ot/localJarClasspathSnapshot.lock
#	.idea/workspace.xml

+++ app/src/main/java/com/att/kiwilauncher/util/CheckLink.java
@@ -21,7 +21,7 @@ public class CheckLink {
         return check;
     }
 
-    public String stringForTime(long timeMs) {
+    public String stringForTime(int timeMs) {
       StringBuilder formatBuilder = new StringBuilder();
         Formatter formatter=new Formatter(formatBuilder, Locale.getDefault());;
         if (timeMs == C.TIME_UNSET) {
